Reset current depth when encountering { at BOL
Previously this also required the nesting level to be zero. However,
if the nesting level was incorrect, this prevented the nesting level
from getting reset increasing the error's effect. For example, when
the program processed hundreds of concatenated C files, the nesting
level at some point reached 17703, and performance dropped dramatically.